fix pages file structure and remove root
