+++ client/src/pages/BrowsePage.jsx
@@ -1,0 +1,3 @@
+export default function BrowsePage() {
+    return "Hello, I am the BROWSE Page";
+}
