--- client/src/pages/BrowsePage/BrowsePage.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { makeStyles, createStyles } from '@material-ui/core/styles';
-
-export default function BrowsePage() {
-    return "Hello, I am the BROWSE Page";
-}
