--- client/src/pages/BrowsePage/index.js
@@ -1,3 +1,0 @@
-import BrowsePage from 'pages/BrowsePage/BrowsePage';
-
-export default BrowsePage;
