+++ client/src/pages/FavoritesPage.jsx
@@ -1,0 +1,3 @@
+export default function FavoritesPage() {
+    return "Hello I am the FAVORITES Page";
+}
