--- client/src/pages/FavoritesPage/FavoritesPage.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import { makeStyles, createStyles } from '@material-ui/core/styles';
-
-
-export default function FavoritesPage() {
-    return "Hello I am the FAVORITES Page";
-}
