--- client/src/pages/FavoritesPage/index.js
@@ -1,3 +1,0 @@
-import FavoritesPage from 'pages/FavoritesPage/FavoritesPage';
-
-export default FavoritesPage;
