--- client/src/pages/LandingPage/index.js
@@ -1,3 +1,0 @@
-import LandingPage from 'pages/LandingPage/LandingPage';
-
-export default LandingPage;
