--- client/src/pages/NewsPage/index.js
@@ -1,3 +1,0 @@
-import NewsPage from 'pages/NewsPage/NewsPage';
-
-export default NewsPage;
