--- client/src/pages/Root/Root.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import Header from '../../components/Header'
-
-// Todo: set this up to pass in currently selected page
-
-export default function Root() {
-    return (
-        <>
-            <Header isLoggedIn={false}/>         
-        </>
-        
-    );
-}
