--- client/src/pages/Root/index.js
@@ -1,3 +1,0 @@
-import Root from 'pages/Root/Root';
-
-export default Root;
