+++ client/src/pages/index.js
@@ -1,0 +1,4 @@
+export { default as BrowsePage } from './BrowsePage';
+export { default as FavoritesPage } from './FavoritesPage';
+export { default as LandingPage } from './LandingPage';
+export { default as NewsPage } from './NewsPage';
